Move classes to their own file
