+++ features/mpabout/src/main/java/com/jpp/mpabout/AboutFragment.kt
@@ -15,12 +15,12 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.jpp.mp.common.extensions.*
-import com.jpp.mp.common.fragments.MPFragment
+import com.jpp.mp.common.extensions.cleanView
+import com.jpp.mp.common.extensions.send
+import com.jpp.mp.common.extensions.setScreenTitle
+import com.jpp.mp.common.extensions.web
 import com.jpp.mp.common.viewmodel.MPGenericSavedStateViewModelFactory
 import com.jpp.mpabout.databinding.FragmentAboutBinding
-import com.jpp.mpabout.databinding.ListItemAboutBinding
 import com.jpp.mpdesign.ext.getColor
 import dagger.android.support.AndroidSupportInjection
 import kotlinx.android.synthetic.main.fragment_about.*
@@ -107,33 +107,6 @@ class AboutFragment : Fragment() {
                 setStartAnimations(it, R.anim.activity_enter_transition, R.anim.activity_exit_transition)
                 setExitAnimations(it, R.anim.activity_enter_transition, R.anim.activity_exit_transition)
             }.build().launchUrl(it, Uri.parse(uriString))
-        }
-    }
-
-    class AboutItemsAdapter(private val items: List<AboutItem>, private val itemSelectionListener: (AboutItem) -> Unit) : RecyclerView.Adapter<AboutItemsAdapter.ViewHolder>() {
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-            return ViewHolder(
-                    DataBindingUtil.inflate(
-                            LayoutInflater.from(parent.context),
-                            R.layout.list_item_about,
-                            parent,
-                            false
-                    )
-            )
-        }
-
-        override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            holder.bindItem(items[position], itemSelectionListener)
-        }
-
-        override fun getItemCount(): Int = items.size
-
-        class ViewHolder(private val itemBinding: ListItemAboutBinding) : RecyclerView.ViewHolder(itemBinding.root) {
-            fun bindItem(item: AboutItem, selectionListener: (AboutItem) -> Unit) {
-                itemBinding.viewState = item
-                itemBinding.executePendingBindings()
-                itemView.setOnClickListener { selectionListener(item) }
-            }
         }
     }
 }
